treadmill-ci.yml: update netboot-raspberrypi-nbd image id
This new image incorporates an API change in the control-socket
interface between puppet binary (baked into the image) and the
supervisor (running on the top-of-rack server). This new puppet
version will create a /run/tml/host-id file, but otherwise should not
have any visible interface changes to the GitHub actions workflows.